requirements.txt
Flask
python-nbt
Werkzeug